Hoist form to encompass both overrides and GPP
